Refactored into controller and route files and style fixes
